StreamID => long

+++ src/System.Net.Quic/ref/System.Net.Quic.cs
@@ -43,7 +43,7 @@ namespace System.Net.Quic
         public override void Flush() => throw null;
         public override int Read(byte[] buffer, int offset, int count) => throw null;
         public override void Write(byte[] buffer, int offset, int count) => throw null;
-        public int StreamId => throw null;
+        public long StreamId => throw null;
         public void ShutdownRead() => throw null;
         public void ShutdownWrite() => throw null;
     }
